исправил по замечаниям
