+++ src/App.js
@@ -7,13 +7,19 @@ export const App = () => {
   return (
     <div className="wrapper">
       <div className="input-space">
-        <Input>Дата и время заказа</Input>
-        <Input placeholder="ММ.ДД.ГГГГ" value="06.12.2021" incorrect>
-          Дата и время заказа
-        </Input>
-        <Input placeholder="ММ.ДД.ГГГГ" value="06.12.2021" disabled>
-          Дата и время заказа
-        </Input>
+        <Input label="Дата и время заказа" />
+        <Input
+          label="Дата и время заказа"
+          placeholder="ММ.ДД.ГГГГ"
+          defaultValue="06.12.2021"
+          incorrect
+        />
+        <Input
+          label="Дата и время заказа"
+          placeholder="ММ.ДД.ГГГГ"
+          defaultValue="06.12.2021"
+          disabled
+        />
       </div>
     </div>
   );
